altera: platform drivers update
Split drivers to correspond to current structure.
